Moved existing core scripts into subfolders
