+++ Assets/Scripts/Core/Messages/MessageManager.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Akashic.Core
+namespace Akashic.Core.Messages
 {
     internal sealed class MessageManager
     {
