+++ Assets/Scripts/Runtime/Controllers/OptionsMenu/Messages.cs
@@ -1,4 +1,4 @@
-using Akashic.Core;
+using Akashic.Core.Messages;
 
 namespace Akashic.Runtime.Controllers.OptionsMenu
 {
