+++ Assets/Scripts/Runtime/MonoSystems/Debugger/IDebuggerMonoSystem.cs
@@ -1,4 +1,4 @@
-using Akashic.Core;
+using Akashic.Core.MonoSystems;
 
 namespace Akashic.Runtime.MonoSystems.Debugger
 {
