+++ Assets/Scripts/Runtime/MonoSystems/ExplorationState/IExplorationStateMonoSystem.cs
@@ -1,4 +1,5 @@
 using Akashic.Core;
+using Akashic.Core.MonoSystems;
 
 namespace Akashic.Runtime.MonoSystems.ExplorationStates
 {
