+++ Assets/Scripts/Runtime/MonoSystems/ExplorationState/Messages.cs
@@ -1,5 +1,4 @@
-using UnityEngine;
-using Akashic.Core;
+using Akashic.Core.Messages;
 
 namespace Akashic.Runtime.MonoSystems.ExplorationStates
 {
