+++ Assets/Scripts/Runtime/MonoSystems/GameStates/Messages.cs
@@ -1,4 +1,4 @@
-using Akashic.Core;
+using Akashic.Core.Messages;
 
 namespace Akashic.Runtime.MonoSystems.GameStates
 {
