+++ Assets/Scripts/Runtime/MonoSystems/Party/IPartyMonoSystem.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using Akashic.Core;
+using Akashic.Core.MonoSystems;
 using Akashic.Runtime.Controllers.PartyMemberBattle;
 
 namespace Akashic.Runtime.MonoSystems.Party
