+++ Assets/Scripts/Runtime/MonoSystems/PlayerPrefs/IPlayerPreferencesMonoSystem.cs
@@ -1,4 +1,5 @@
 using Akashic.Core;
+using Akashic.Core.MonoSystems;
 
 namespace Akashic.Runtime.MonoSystems.PlayerPrefs
 {
