+++ Assets/Scripts/Runtime/MonoSystems/PlayerPrefs/Messages.cs
@@ -1,4 +1,4 @@
-using Akashic.Core;
+using Akashic.Core.Messages;
 
 namespace Akashic.Runtime.MonoSystems.PlayerPrefs
 {
