+++ Assets/Scripts/Runtime/MonoSystems/Scene/ISceneMonoSystem.cs
@@ -1,4 +1,5 @@
 using Akashic.Core;
+using Akashic.Core.MonoSystems;
 using UnityEditor;
 
 namespace Akashic.Runtime.MonoSystems.Scene
