+++ Assets/Scripts/Runtime/MonoSystems/Story/IStoryMonoSystem.cs
@@ -1,4 +1,4 @@
-using Akashic.Core;
+using Akashic.Core.MonoSystems;
 
 namespace Akashic.Runtime.MonoSystems.Story
 {
